[Silver III] Title: 주유소, Time: 368 ms, Memory: 35740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -22,10 +22,10 @@ public class Main {
         }
         br.close();
 
-        int resultCost = 0; // 출력될 최소 비용
+        long resultCost = 0; // 출력될 최소 비용
         int i = 0;
         while(i < GAS_ARR.length - 1){
-            int num = ROAD_ARR[i]; // 첫 번째 도로 길이 세팅
+            long num = ROAD_ARR[i]; // 첫 번째 도로 길이 세팅
             int nextIdx = i+1;
             while(GAS_ARR[i] <= GAS_ARR[nextIdx] && nextIdx < ROAD_ARR.length) { // 현재 도시의 주유값이 다음 도시의 주유값보다 싸거나 같으면
                 num += ROAD_ARR[nextIdx];   // 다음 도로의 길이를 합함 (싼 비용으로 계산하기 위함)
